fixed sierpinski simultaneous subdivision and consecutivity constraint
If the sierpinski curve crosses an edge which is to be subdivided then both triangles attached to that edge should be subdivided at the same time.
This is where the consecutivity constraint comes from.
A couple of the main functions of SierpinskiFill had to be totally revamped to be able to handle multiple nodes.